requirements.txt
# Core dependencies
python-telegram-bot>=20.7
pandas>=2.2.3
numpy>=2.0.0
Pillow>=10.0.0
requests>=2.31.0

# Vector search and ML
faiss-cpu>=1.7.4
torch>=2.2.0
transformers>=4.37.0

# Logging
python-json-logger>=2.0.7

# Image processing and computer vision
opencv-python-headless>=4.8.0
scikit-image>=0.21.0

# Machine learning and neural networks
torchvision>=0.15.2
onnxruntime>=1.15.1
onnx>=1.14.0

# Storage and caching
redis>=4.6.0
pymongo>=4.5.0

# Monitoring
colorama>=0.4.6
psutil>=5.9.0

# Async operations
aiohttp>=3.8.5

# Web and API
fastapi>=0.103.0
uvicorn>=0.23.0

# Development tools
pytest>=7.4.0
tqdm>=4.66.0

# Docker dependencies
gunicorn>=21.2.0

# Additional dependencies
python-dotenv>=1.0.0
cryptography>=41.0.0
pyTelegramBotAPI>=4.14.0
scikit-learn>=1.3.0

# YOLO and object detection
ultralytics>=8.0.0

# Analytics and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
